requirements.txt
streamlit == 1.31.0
nltk == 3.8.1
streamlit-extras == 0.4.0
scikit-learn == 1.4.1.post1
fastapi == 0.109.2
uvicorn == 0.27.1
python-dotenv == 1.0.1
numpy == 1.26.4